Reduce counter to int from long

+++ app/models/IndividualCheckCount.scala
@@ -18,7 +18,7 @@ package models
 
 import play.api.libs.json._
 
-case class IndividualCheckCount(_id: String, attempts: Long)
+case class IndividualCheckCount(_id: String, attempts: Int)
 
 object IndividualCheckCount {
   implicit val format = Json.format[IndividualCheckCount]
